refactor: DTO에서 request/response를 제외한 값 객체들을 분리
